Fix memory recall bug.
Knowledge consolidation (this bit is purely for me so i can explain and consolidate on the lesson learnt in pointer arithmatic):
When incrementing through addresses the stride each address is incremented by is determined by the datatype of the address.
Initially i was incrementing through the uint128 datatype to try output each byte and each address being given to me was 8bytes apart,
this was because i was incrementing from an unsigned long memory address (which in my system is 8bytes in size). A datatype that is
1byte large in C is a char, therefore in order to correctly increment byte-by-byte through all 16bytes of the uint128, i must first
cast the address into a char and then increment.

This has caused quite a kerfuffle so glad that has been solved.